requirements.txt
# Core
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0

# NLP
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1
bertopic==0.16.0
umap-learn==0.5.5
hdbscan==0.8.33
nltk==3.8.1

# Finance data
yfinance==0.2.36
pandas-datareader==0.10.0

# News/text data
newsapi-python==0.2.7
praw==7.7.1

# Modeling
xgboost==2.0.3
imbalanced-learn==0.11.0

# Evaluation & visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0
shap==0.44.1

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
requests==2.31.0